DB, Logging, Routing
Aside from the super basic setup, we now have some other basic items in
place such as a DB Pool injected into the middleware as well as a
middle-ware logging framework and a log provider. We've also setup some
basic routing.

This concludes the minimum set of requirements (mostly) to start
building out the API. Next order of business is to handle the
authentication and session management.